Modify R200 launch file.
Modified the R200 3D sensor launch file
to conform more closely to the other
open_ni2-style sensor launch files. This
allows the data processing arguments passed
in to correctly pass through and control
image processing.

This allows the turtlebot_follower package,
for example, to set "depth_processing" to
true to enable the /camera/depth/depth_rect/
topic with the R200 as it can with other
cameras.